gmail/requirements.txt
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dateutil==2.9.0.post0
requests==2.32.3

